+++ selendroid-test-app/src/test/java/io/selendroid/driver/SelendroidUnknownCommandHandlingTest.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.selendroid.nativetests;
+package io.selendroid.driver;
 
 import io.selendroid.support.BaseAndroidTest;